Move files to weapon folder
