+++ src/data/weapons/handguns.js
@@ -1,4 +1,4 @@
-import defaultHandgunsProgress from '../defaults/progress/handguns'
+import defaultHandgunsProgress from '../defaults/progress/weapons/handguns'
 
 const base = ['P890', '.50 GS', 'X12', 'Basilisk', 'X13 Auto']
